fix(guided remediation): reduce memory footprint by computing dependency subgraphs instead of chains (#1538)
Guided remediation had been using `DependencyChains` to track paths to a
vulnerable package (for computing things like depth and which direct
dependencies to relax). It was computing *every* possible path in the
graph to a dependency, which grows roughly exponentially with depth /
connectivity. This was using an unreasonable amount of memory on some
particularly large/complex projects.

I've changed the logic to instead compute one `DependencySubgraph` - the
set of nodes and edges that would contain every path to a dependency.
This should significantly reduce memory usage (and cpu usage from
allocs) when running on larger projects.

This change has touched quite a few places in the code, and the logic is
a bit complex. I've tried my best to check that everything still behaves
as expected.