feat: Container json testing (#1604)
Fork of #1602 

Adds a JSON matching method, and implements a path enumeration function
to help mock values in the snapshot to avoid polluting the logs with
very long JSON values that we are not necessarily testing.

Moved the normalisation step into the match snapshot function itself to
ensure that it is always the **last** step taken.