test: reduce noise in cmd cases using a custom `slog` handler (#1631)
Currently we got a lot of noise from `osv-scalibr` libraries which makes
it hard to see what actually failed - since all of these are going
through the `slog`, we can use our own custom handler to drop records
with specific prefixes.

I believe there are some talks happening with the `osv-scalibr` team to
see if some of these can be toned back, and also I think we will
probably want to capture some of these at some point as part of valid
tests, but short-term this should help make things more manageable.

I'm opening this as a draft as I don't know if we do actually want to
land this, in part because I think this is going to almost immediately
break as we've switching to `slog` ourselves which involves using a
custom `slog` handler meaning this one will get overridden, but I still
wanted to share the implementation either way as folks might like to
cherry-pick it locally depending on what they're working on 🙂