feat: remove "verbose" verbosity level (#1686)
We're only actually using this level in one place currently, and it is
not one of the log levels defined by `slog` by default - I also think
there's an argument to be made that "verbose" itself is something of a
mode rather than a level; i.e. it implies maybe all logs should be shown
regardless of their level, or debug, or some other combination.

Either way I think we should remove it as part of v2 while we can to
make things a bit simpler and we can introduce it back in a minor
version if/when we decide to use it.