feat: replace custom `reporter` package with `slog` (#1666)
Our custom `reporter` package has served us well, but it's annoying to
have to pass it through to every function especially when you're using
the libraries, so this switches us over to using `slog` which uses the
global default logger - the notable downside with this is as a global
tests that depend on logging output cannot be run in parallel, but that
shouldn't be a big deal for downstream consumers as they should be
depending on the structured data returned by package functions (i.e. our
packages only log information that could be useful for a human to debug
stuff).

This downside does impact our CLI test suite since that currently _does_
depend on that logging output as part of feedback - for now I've just
"fixed" this by removing `t.Parallel` from our CLI tests, which makes
them take a lot longer but should make this landable straightaway, after
which we'll be free to chop and change what we like since we'll have
gotten rid of the public `reporter` package and log output is not
covered by semver.

I've already got a few ideas following on from this on how we might be
able to speed our CLI tests back up, starting with attempting to chop
the tests up into actual small packages, since Go supports running
package tests in parallel regardless of their dependency on globals.

---

I do still need to give this a manual review, but I believe the primary
thing left before this could be landed (warts or not) is to ensure when
we're outputting JSON all other output goes to `stderr`