refactor: move cli building into a dedicated internal package to enable relocating tests (#1737)
As we've switched to using a global logger, we can no longer run tests
that involve checking that output in parallel which greatly impacts our
cmd tests since that's all about what we output - however, `go test`
always runs tests across packages in parallel, meaning we can improve
our test suite performance by refactoring so that our cmd tests are
spread across multiple (sub)packages, which this does.

The key change to enabling this is moving the build-and-run of `cli.App`
from the `main` package in `cmd/osv-scanner` into a dedicated internal
package, since that results in an internally public function we can
import elsewhere from other packages including tests under the guise of
testing the specific command each subpackage represents.

Overall for me locally this takes `TEST_ACCEPTANCE=true go
./cmd/osv-scanner/... -count 1` from ~6 minutes to ~3 minutes, and also
should hopefully mean caching happens slightly more frequently (though
it's still unlikely to be great).

While I still have some reservations about this, overall I feel like its
a good move due to a couple of small bonus positives such as now having
the tests in their own `_test` packages and being more explicitly tied
with the public `Command` of each package.

I think my remaining reservations are around how unlike other unitish
tests this blurs the line on what is actually being tested since each
command package is really still testing the full `cli.App` with all
commands mounted, meaning behaviour unrelated to the command/package can
be inadvertently tested - that's not the end of the world, but ideally
it would be nice if we could untangle things some more.

We already see that a bit with the remaining tests in
`cmd/osv-scanner/main_test.go` which are arguably testing
`cmd/osv-scanner/internal/cmd#Run` and so could live there, but they're
calling `testcmd.Run` - we could alternatively call
`cmd/osv-scanner#run` but then that'd ideally mean replicating some of
the logic in `testcmd`...

A change we could do that might help with this is switch to passing
`cmd.Run` the array of commands to mount, meaning each package could be
more explicitly based on testing their particular thing and that would
also mean improve cache reuse a little (since right now a change to any
command means the tests for all command packages are technically
changed), but then we'd arguably be testing less and in theory increase
the chance of missing bugs related to the interactions between different
commands (though that might be a thing at all 🤷).

Ultimately, this does improve the test suite performance and its all
internal stuff so it's probably fine (and maybe best) to aim at landing
it as-is to start with, and then we can follow-up with changes later as
our feelings come out in the wash.

Finally, there is some further refactors with the test suite that I'm
going to look into as a follow-up such as possibly moving the fixtures
into the subpackages and deduplicating the methods in `testcmd` as I
think we can get away with just the `testcmd.Run` method with a few
tweaks.