feat(xml): keep the white space between attributes (#1805)
For start elements that have attributes spanning multiple lines, the
default behavior of `encoding/xml` is to encode the element into one
single line, which results a lot diffs.

To keep the "formatting" of attributes, this PR adds a `Diff` field to
`Attribute` that represents the diff whitespace between two attributes.

This PR also renames `Space` in `StartElement` to `Diff` to have similar
indication on keeping the whitespace.