feat(xml): store the original text in CharData (#1825)
We can not fully rely on `escape` flag to decide whether to escape texts
in CharData or not since there can be both escaped and unescaped texts.

When decoding a token, escape sequences are replaced by the characters
that they represent. To be able to write the original text, we add the
slice of bytes to CharData to represent the texts before any
replacement. And when encoding a token, we use the original bytes to
minimize the diff.