feat: adding stable tag to container images in release workflow and g… (#1882)
# Description

This Pull Request resolves #1563 by ensuring that the `stable` Docker
tag is only created for properly versioned tags following Semantic
Versioning in the format `MAJOR.MINOR.PATCH` (e.g., `2.0.0`). Tags such
as `v2.0.0`, `2.0.0-beta`, or any non-compliant version will not be
considered stable.

# Changes

- Added step in `goreleaser.yml` (GitHub Actions release workflow) to
detect whether the current tag is a **stable semantic version** with the
help of regex, the result is stored in the `STABLE` environment
variable.
- In `.goreleaser.yml` file, added `docker_manifests` entry for the
`stable` tag and used `skip_push` to skip pushing docker manifests in
amd64 and arm64 of osv-scanner image when the `STABLE` environment
variable is false.