fix: also compare PURL when sorting inventories (#1871)
https://github.com/google/osv-scanner/issues/1865

Some inventories may have the same name and version but refer to
different packages (e.g. Maven packages with the same artifact IDs).
They may be considered as the same with the existing compare function
and deduped during scanning. This PR adds PURL comparison so that these
packages are not deduped unexpectedly.