test: Don't parallel the one test that can't be parallel (#1907)
Fix the version and help output not being thread safe by avoiding
parallel for that test.

Reasoning:
cli.VersionPrinter is a global variable that gets overwritten every .Run
call with a new function. I'm still not too clear on why that would
cause the logs to be messed up though.