ci: improve markdown link checker setup (#1892)
Currently its very easy for broken links to not be noticed as scheduled
workflows don't really have a good way of signaling when they've failed
since people can have different notification settings and the actual
subject of the notification for a failed workflow I believe is also
based on who actually last modified the workflow.

To address that I've added the full codebase link checker as part of our
prerelease steps and also had the workflow run on pushes and pull
requests that modify markdown files, though note those are still running
against the modified files.

While doing this I also merged our two workflows into one to make it
easier to maintain and switched us to using a fork of the original
action as that has been deprecated in favor of a new tool whose action
is using another linting/reporting tool that I don't think we want to
use (reviewdog)