chore: Update osv-scalibr (#1935)
Update osv-scalibr to the latest version to unblock various osv-scanner
feature developments, including:
- OSV-Scanner/OSV-Scalibr convergence #1769
- Guided Remediation to use scalibr migrated code.

OSV-Scalibr changed the package format to use a combination of PURLType
and Plugin Name string rather than a pointer to the extractor itself.
This means to get ecosystem we need to set PURLType and Metadata in our
mocks.