feat: Use OSV-Scalibr scan function directly (#1936)
Closes #1769 

Behaviour changes / notes:
- OSV-Scalibr's Scan() function does DFS rather than BFS, so the order
of files being read in the snapshot changes slightly.
- Their Scan() function also does not deduplicate the output of the
extractor, so the SBOM with duplicates will return double the results. I
think it's fine to have that behavior change for now unless someone
complains about it.
- Unfortunately, we also double log the path to the file in the error,
because both the extractor and the walker both prepend the path the
error originated from. This is a TODO in osv-scalibr to remove input
path references in the extractor errors.
(https://github.com/google/osv-scalibr/issues/821)

---------

Co-authored-by: Mend Renovate <bot@renovateapp.com>