feat: deprecate `--sbom` in favor of just using `-L` (#1961)
I think the only difference between the two flags is that `--sbom` will
give you an extra message if you provide a file which we fail to make to
an extractor we give you an extra message about making sure to check the
filename, which I actually think is arguably better with `-L` because
imo you can't reasonably assume the scanner will be able to magically
figure out e.g. `my.awesome.file.json` is meant to be an SBOM unlike
with the current flag?

This is the first of part #1960