feat: Add CDX Location field to the output source string (#2038)
The CDX extractor now will output the occurrence location if it found it
in the SBOM. This just adds that to the output.

- Also added a new test for it
- ~This is running into a windows issue, as does not understand that
absolute paths don't need to be joined to absolute paths.~~ Temporarily
fixed, long term fix involves saving cdx Location in metadata.

~~Turns out this doesn't quite work yet, as scalibr sorts locations. We
got to fix that first.~~ Fixed

Closes #1972