feat: support running detectors (#2021)
This adds experimental support for having the scanner run
[detectors](https://github.com/google/osv-scalibr/blob/main/detector/detector.go)
which are plugins aimed at identifying vulnerabilities through more
interactive means than extractors such as by reading config files or
making HTTP requests with crafted payloads.

Currently this does not have a lot of tests as only a handful of
detectors exist and most are based on very specific situations which
cannot be changed without shipping support for custom configuration
(which is its own whole thing), but I have managed to at least get one
with a positive finding using the `weakcredentials/etcshadow` detector
scanning a custom image.

Also this does not remove the requirement of at least one extractor
being enabled and a package found since that too would be a much larger
change and with the current pool of detectors available being so limited
I don't think it's appropriate to drop that requirement just yet.

Resolves #1969