fix: invalid ecosystem error (#2156)
Fixes: #2151

It looks like osv.dev will return invalid ecosystem for all queries in a
batchquery if 1 query has an invalid ecosystem. This means we have to
add a check in osv-scanner to prevent any invalid ecosystems from being
sent.

This does mean if we release osv-scanner with extractors to support
future ecosystems that are currently invalid, people will have to update
to get support for those ecosystems when supported in osv.dev in the
future, instead of being able to use existing version of osv-scanner. I
don't think this is a common case, so this is acceptable.

Changes in this PR:
- Added Podfile.lock to test cases, should return 127 if scanned alone
- Excude Podfile.lock and Package.resolved because we don't support the
Swift Cocoapods ecosystem (SwiftURL is for the Swift package manager,
which is different from Cocapods?)
- Added a IsValid() check for ecosystems, which contains all the current
ecosystems within osvschema. Missing ecosystems would be caught by the
exhaustive linter in the future.
- Filter on IsValid ecosystems so we don't send queries with invalid
ecosystems.
- Fix what seemed to be a bug with unknown Maven packages not being
skipped.