feat: Update scalibr to the latest version (#2284)
Update osv-scalibr to the latest version. This required a few changes:

- Remove baseimagematch and use the base image enricher in osv-scalibr
- Add enrichers to the preset.go
- Custom clients for enrichers to set user agent. Create a insight
client helper function for v3alpha


I'm not sure what the snapshot update is for, and how we are getting new
severity when we weren't before though.

Enrichers might be making network requests during initialisation (I need
to look into how grpc initialisation works). This will be a problem if
--offline is set to true.

---------

Co-authored-by: osv-robot <osv-robot@google.com>