fix: improve logic for `--all-packages` flag  (#2311)
Update logic so that unscannable packages are still filtered for
vulnerabilities and license matching, but they will be re-added to the
final JSON result if the `--all-packages` flag is passed in.

Also update docs to clarify the order of precedence between
`--all-packages` flag and `PackageOverrides` actions in the config file.

Fixes #2157