feat: replace `requirementsenhanceable` extractor with transitive enricher (#2294)
On the surface this was pretty much a drop-in replacement which is
really exciting, but there are a couple of quirks that we may or may not
want to deal with before landing this.

The first one is very obvious which is that our "found packages" count
is very different - it's technically correct, but I think ideally it
would be good if we could express that we found x more packages via the
enricher.

The second one is more interesting and I suspect one we'll probably
ignore for the time being, but right now `scalibr` implicitly adds
extractors that are required by enrichers as part of doing the scan
which is at odds with our (experimental) abilities to enable and disable
plugins - for now I've made it so that the enricher is only enabled if
the `requirements` extractor is enabled, since we don't have that many
enrichers with plugin requirements right now.

Resolves #2289